requirements.txt
gradio
python-dotenv
groq
langchain
langchain-core
langchain-community
sentence-transformers
chromadb
PyPDF2
transformers
torch
torchaudio
SpeechRecognition
pydub
ffmpeg-python
gTTS
elevenlabs
faiss-cpu
requests
numpy
typing-inspect
typing_extensions
pypdf